+++ nn/loss/softmax.go
@@ -20,5 +20,5 @@ func (*Softmax) Loss(predict, targets *tensor.Tensor) *tensor.Tensor {
 	exps := predict.Sub(max).Exp()
 	sum := exps.SumAxis(1)
 	logSoftmax := predict.Sub(max).Sub(sum.Log())
-	return logSoftmax.MulElem(targets).Scale(-1).Sum().Scale(1 / float64(rows))
+	return logSoftmax.MulElem(targets).Scale(-1).Sum().Scale(1 / float32(rows))
 }
