+++ nn/tensor/utils.go
@@ -1,7 +1,7 @@
 package tensor
 
 func Ones(rows, cols int) *Tensor {
-	data := make([]float64, rows*cols)
+	data := make([]float32, rows*cols)
 	for i := range data {
 		data[i] = 1
 	}
@@ -9,12 +9,12 @@ func Ones(rows, cols int) *Tensor {
 }
 
 func Zeros(rows, cols int) *Tensor {
-	data := make([]float64, rows*cols)
+	data := make([]float32, rows*cols)
 	return New(data, rows, cols)
 }
 
-func Numbers(rows, cols int, n float64) *Tensor {
-	data := make([]float64, rows*cols)
+func Numbers(rows, cols int, n float32) *Tensor {
+	data := make([]float32, rows*cols)
 	for i := range data {
 		data[i] = n
 	}
